[create]-Function Make Menu Links
Description:
Function Make Menu Links:
In this function, it puts them into one presentation by taking two inputs
The Code of a Loop to Create Four Links. In Each It of the Loop, a new link element is creating the JavaScript createElement Function. The link element is then added to the webpage, Specification to the Rightmenu. The Text Content of the Link is Set to One of the Input Values, in Order. So, the first link will show the value of home, the second link will show of requests, the first link will show the value of requestshop, and the first link will show the value of call.