[create]-base structure for headContent
Description:
This is a sketch-like sketch for the headContent section.
It has basic styles and the font has not been given to it yet, and it has been added here only to place and specify the display and performance of each demon in this section. In the next parts, small and large edits will be made in this part and the final result document will be written in each part.