[add]-Engine animation
Description:
Animation of the motor movement is based on the position of the screen. That is, when the user reaches that level, the engine starts to move. We were able to do this by getting the height of the page and the definition of an animation for the object in question.
We also used scroll to scroll.