[create]-mobile safir menu
Description:
This is the code for the ambassador page. This is the mobile menu code. We wrote this modular code and added it to the main Script page with Import. This code is manually adjusted and can be manually adjusted to the menu.
And adjusted with the structure of the case. This is just a base structure and changes will apply later.